+++ pages/superstar-sites.tsx
@@ -25,8 +25,8 @@ import { styled } from "@mui/system";
 import LayoutContainer from "components/LayoutContainer";
 import StyledHeader from "components/StyledHeader";
 import useValidateUserToken from "hooks/useValidateUserToken";
-import { handleWPLogin } from "./handle-wp-login";
-import { colors } from "./colors";
+import { handleWPLogin } from "../utils/handle-wp-login";
+import { colors } from "../utils/colors";
 
 interface SuperstarSite {
   id: number;
